Add files via upload
I have an error with the infinite loop part, but I think I know how to solve it but don't want to mess anything up. I think that I need to move the code that sets up the random number into Game1 file so that each time a new game is created (of Game1), then it will also create a new random number.